MY FIRST PERIODIC TASK (LAB 3-2)
We have changed the Init task of the Rtems System so that it creates a
housekeeping task and then it eliminates itself. The housekeeping task
is a periodic task in charge of controlling the state of the system. To
simulate the reading of different sensors and the generation of
telemetries with the parametres, the task gets random values using the
clock of the system and prints them continiously.
In order to add a task and use the clock, we have needed to change some
configuration macros in the corresponding file (rtems_config.h).