CONTINUATION OF THE PERIODIC TASK(LAB 3-2)
Firstly we have set the task period to one second by using the RTEMS
primitives rtems_task_wake_after() and
rtems_clock_get_ticks_per_second().

Secondly we have created a new source file (housekeeping.c) in order to
create the functions related to housekeeping task operations.
This file contains two functions:
-do_housekeeping saves random values (using rand_r(&seed) in a vector,
which emulates the collection of measurements. It also prints the values
of the array, which emulates the generation of telemetries, when it
elapses a certain number of activations. This function is called inside
the loop of the houskeeping task.
-init_housekeeping just initialize the value of the seed used to get
random values to 0. This function is only called once in the main
function of the housekeeping task.

Thirdly we have created a new source file (delay.c). The function
task_delay_until() allow us to avoid the drift caused by the generation
of the telemetry by establishing the delay as a number of ticks since
the system boot.