LAB 3-4
We have modified the application to handle the received telecommands and
generate the corresponding telemetries.
The telemetries are sent though the UART interface.
We have defined the telemetry channel and we have built the telemetry
packets. Finally the manager and housekeeping tasks have been changed to
handle the telecommands and generate the telemetries.