LAB 3-5
In this practical we have adapted our code in order to receive
telecommands via the UART.
To this end we have created a character FIFO and the functions to manage
it. They are declarated in the header file uart_rx_queue.h, and
implemented in the source file uart_rx_queue.c.
In the second place, we have implemented the UART driver following the
guidelines set by the RTEMS device-independent software layer. Different
primitives have been declared in driver.h and implemented in driver.c:
riscv_uart_driver_initialize(), riscv_uart_driver_open(),
riscv_uart_driver_write(), riscv_uart_driver_read(),
riscv_uart_driver_close().
Finally, it has been necessary to adapt the applications SW: we have
registered the driver in the rtems configuration file; we have udated
the telemetry channel to use the POSIX API syscalls; we have created the
tc_rx_task, wich substitutes the emu_tc_rx_task and is in charge of
receiving the telecommands through the uart.